refactor: move attribute ordering related components to separate file
